fix: refactor tests
